Add comment about caching location.
See https://stackoverflow.com/questions/30104047/how-can-i-decorate-an-instance-method-with-a-decorator-class

Since the method is wrapped on import rather than on instance creation,
if used as a decorator, cache_value will save the cache on a per class
basis rather than a per instance basis.

This means that if used to decorate a method of a class, a call to the
method from one instance will affect the cache of all instances.